Added tests
Added some more tests
